backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
pyjwt
werkzeug
